requirements.txt
# Plateforme Infographie Intelligente - Dépendances
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Extraction documents
pypdf==4.0.1
python-docx==1.1.0
python-pptx==0.6.23

# Analyse de contenu (optionnel: OpenAI pour meilleure extraction)
openai==1.12.0
spacy==3.7.2  # NLP local performant
# Modèle fr_core_news_sm à télécharger via: python -m spacy download fr_core_news_sm

# Génération visuelle (Pillow >=10.2 pour éviter erreur de build sur Windows)
Pillow>=10.2.0,<11
jinja2==3.1.3
weasyprint==60.2

# Utilitaires
pydantic==2.5.3
pydantic-settings==2.1.0
python-dotenv==1.0.0
regex>=2023.0.0
